rename: authentication module & subsidaries to login
