+++ packages/client/lib/app/app_module.dart
@@ -2,7 +2,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:nokhte/app/modules/irl_nokhte_session/irl_nokhte_session_module.dart';
 import 'package:nokhte/app/modules/storage/storage_module.dart';
 import 'core/modules/supabase/supabase_module.dart';
-import 'modules/authentication/authentication_module.dart';
+import 'modules/login/login.dart';
 import 'modules/collaboration/collaboration_module.dart';
 import 'modules/home/home_module.dart';
 
@@ -14,7 +14,7 @@ class AppModule extends Module {
 
   @override
   void routes(r) {
-    r.module('/', module: AuthenticationModule());
+    r.module('/', module: LoginModule());
     r.module('/home/', module: HomeModule());
     r.module('/collaboration/', module: CollaborationModule());
     r.module('/irl_nokhte_session/', module: IrlNokhteSessionModule());
