--- packages/client/lib/app/modules/authentication/data/contracts/authentication_contract_impl.dart
@@ -1,74 +1,0 @@
-import 'package:dartz/dartz.dart';
-import 'package:nokhte/app/core/constants/failure_constants.dart';
-import 'package:nokhte/app/core/interfaces/logic.dart';
-import 'package:nokhte/app/core/mixins/response_to_status.dart';
-import 'package:nokhte/app/core/modules/user_information/data/data.dart';
-import 'package:nokhte/app/modules/authentication/domain/domain.dart';
-import 'package:nokhte/app/core/error/failure.dart';
-import 'package:nokhte/app/core/network/network_info.dart';
-import 'package:nokhte/app/modules/authentication/data/sources/auth_remote_source.dart';
-
-typedef _AppleOrGoogleChooser = Future<dynamic> Function();
-
-class AuthenticationContractImpl
-    with ResponseToStatus
-    implements AuthenticationContract {
-  final AuthenticationRemoteSource remoteSource;
-  final NetworkInfo networkInfo;
-
-  AuthenticationContractImpl({
-    required this.remoteSource,
-    required this.networkInfo,
-  });
-
-  @override
-  Future<Either<Failure, AuthProviderEntity>> googleSignIn(params) async =>
-      await _signInWith(() => remoteSource.signInWithGoogle());
-
-  @override
-  Future<Either<Failure, AuthProviderEntity>> appleSignIn(params) async =>
-      await _signInWith(() => remoteSource.signInWithApple());
-
-  Future<Either<Failure, AuthProviderEntity>> _signInWith(
-      _AppleOrGoogleChooser getOAuthProvider) async {
-    if (await networkInfo.isConnected) {
-      try {
-        final remoteAuth = await getOAuthProvider();
-        return Right(remoteAuth);
-      } catch (err) {
-        return Left(FailureConstants.authFailure);
-      }
-    } else {
-      return Left(FailureConstants.internetConnectionFailure);
-    }
-  }
-
-  @override
-  AuthStateEntity getAuthState(params) => remoteSource.getAuthState();
-
-  @override
-  addName(NoParams params) async {
-    if (await networkInfo.isConnected) {
-      final res = await remoteSource.addName();
-      return Right(fromSupabase(res));
-    } else {
-      return Left(FailureConstants.internetConnectionFailure);
-    }
-  }
-
-  @override
-  getUserInfo(params) async {
-    if (await networkInfo.isConnected) {
-      final userNamesRes = await remoteSource.getUserInfo();
-      final nokhteSessionsRes = await remoteSource.getFinishedNokhteSessions();
-      return Right(
-        UserJourneyInfoModel.fromSupabase(
-          userNamesRes: userNamesRes,
-          finishedNokhteSessionsRes: nokhteSessionsRes,
-        ),
-      );
-    } else {
-      return Left(FailureConstants.internetConnectionFailure);
-    }
-  }
-}
