--- packages/client/lib/app/modules/authentication/data/data.dart
@@ -1,3 +1,0 @@
-export 'contracts/authentication_contract_impl.dart';
-export 'sources/auth_remote_source.dart';
-export 'models/models.dart';
