--- packages/client/lib/app/modules/authentication/data/models/models.dart
@@ -1,2 +1,0 @@
-export 'auth_provider_model.dart';
-export 'auth_state_model.dart';
