--- packages/client/lib/app/modules/authentication/domain/domain.dart
@@ -1,3 +1,0 @@
-export 'contracts/authentication_contract.dart';
-export 'entities/entities.dart';
-export 'logic/logic.dart';
