--- packages/client/lib/app/modules/authentication/domain/entities/auth_provider_entity.dart
@@ -1,13 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:nokhte/app/core/interfaces/auth_providers.dart';
-
-class AuthProviderEntity extends Equatable {
-  final AuthProvider authProvider;
-  final bool authProviderStatus;
-
-  const AuthProviderEntity(
-      {required this.authProvider, required this.authProviderStatus});
-
-  @override
-  List<Object> get props => [authProvider, authProviderStatus];
-}
