--- packages/client/lib/app/modules/authentication/domain/entities/auth_state_entity.dart
@@ -1,10 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class AuthStateEntity extends Equatable {
-  final Stream<bool> isAuthenticated;
-
-  const AuthStateEntity({required this.isAuthenticated});
-
-  @override
-  List<Object> get props => [isAuthenticated];
-}
