--- packages/client/lib/app/modules/authentication/domain/entities/entities.dart
@@ -1,2 +1,0 @@
-export 'auth_provider_entity.dart';
-export 'auth_state_entity.dart';
