--- packages/client/lib/app/modules/authentication/domain/logic/logic.dart
@@ -1,4 +1,0 @@
-export "add_name.dart";
-export 'get_auth_state.dart';
-export 'sign_in_with_apple.dart';
-export 'sign_in_with_google.dart';
