--- packages/client/lib/app/modules/authentication/domain/logic/sign_in_with_google.dart
@@ -1,16 +1,0 @@
-import 'package:nokhte/app/modules/authentication/domain/domain.dart';
-import 'package:nokhte/app/core/error/failure.dart';
-import 'package:nokhte/app/core/interfaces/logic.dart';
-import 'package:dartz/dartz.dart';
-
-class SignInWithGoogle
-    extends AbstractFutureLogic<AuthProviderEntity, NoParams> {
-  final AuthenticationContract contract;
-
-  SignInWithGoogle({required this.contract});
-
-  @override
-  Future<Either<Failure, AuthProviderEntity>> call(NoParams params) async {
-    return await contract.googleSignIn(params);
-  }
-}
