--- packages/client/lib/app/modules/authentication/presentation/mobx/coordinators/coordinators.dart
@@ -1,2 +1,0 @@
-export './login_screen_coordinator.dart';
-export './login_screen_widgets_coordinator.dart';
