--- packages/client/lib/app/modules/authentication/presentation/mobx/main/main.dart
@@ -1,2 +1,0 @@
-export 'sign_in_with_auth_provider_store.dart';
-export 'get_auth_state_store.dart';
