--- packages/client/lib/app/modules/authentication/presentation/mobx/mobx.dart
@@ -1,2 +1,0 @@
-export 'main/main.dart';
-export './coordinators/coordinators.dart';
