+++ packages/client/lib/app/modules/login/domain/domain.dart
@@ -1,0 +1,3 @@
+export 'contracts/login_contract.dart';
+export 'entities/entities.dart';
+export 'logic/logic.dart';
