+++ packages/client/lib/app/modules/login/domain/entities/auth_provider_entity.dart
@@ -1,0 +1,13 @@
+import 'package:equatable/equatable.dart';
+import 'package:nokhte/app/modules/login/types/types.dart';
+
+class AuthProviderEntity extends Equatable {
+  final AuthProvider authProvider;
+  final bool authProviderStatus;
+
+  const AuthProviderEntity(
+      {required this.authProvider, required this.authProviderStatus});
+
+  @override
+  List<Object> get props => [authProvider, authProviderStatus];
+}
