+++ packages/client/lib/app/modules/login/domain/entities/entities.dart
@@ -1,0 +1,2 @@
+export 'auth_provider_entity.dart';
+export 'login_state_entity.dart';
