+++ packages/client/lib/app/modules/login/domain/entities/login_state_entity.dart
@@ -1,0 +1,10 @@
+import 'package:equatable/equatable.dart';
+
+class LoginStateEntity extends Equatable {
+  final Stream<bool> isAuthenticated;
+
+  const LoginStateEntity({required this.isAuthenticated});
+
+  @override
+  List<Object> get props => [isAuthenticated];
+}
