+++ packages/client/lib/app/modules/login/domain/logic/logic.dart
@@ -1,0 +1,5 @@
+export "add_name.dart";
+export 'auth_get_user_info.dart';
+export 'get_login_state.dart';
+export 'sign_in_with_apple.dart';
+export 'sign_in_with_google.dart';
