+++ packages/client/lib/app/modules/login/domain/logic/sign_in_with_google.dart
@@ -1,0 +1,12 @@
+import 'package:nokhte/app/modules/login/domain/domain.dart';
+import 'package:nokhte/app/core/interfaces/logic.dart';
+
+class SignInWithGoogle
+    extends AbstractFutureLogic<AuthProviderEntity, NoParams> {
+  final LoginContract contract;
+
+  SignInWithGoogle({required this.contract});
+
+  @override
+  call(NoParams params) async => await contract.googleSignIn(params);
+}
