+++ packages/client/lib/app/modules/login/presentation/mobx/coordinators/coordinators.dart
@@ -1,0 +1,2 @@
+export 'login_screen_coordinator.dart';
+export 'login_screen_widgets_coordinator.dart';
