+++ packages/client/lib/app/modules/login/presentation/mobx/main/get_login_state_store.dart
@@ -1,0 +1,23 @@
+// ignore_for_file: library_private_types_in_public_api, must_be_immutable
+
+import 'package:mobx/mobx.dart';
+import 'package:equatable/equatable.dart';
+import 'package:nokhte/app/core/interfaces/logic.dart';
+import 'package:nokhte/app/modules/login/domain/domain.dart';
+part 'get_login_state_store.g.dart';
+
+class GetLoginStateStore = _GetLoginStateStoreBase with _$GetLoginStateStore;
+
+abstract class _GetLoginStateStoreBase extends Equatable with Store {
+  final GetLoginState logic;
+
+  _GetLoginStateStoreBase({
+    required this.logic,
+  }) : authState = logic(NoParams()).isAuthenticated.asBroadcastStream();
+
+  @observable
+  Stream<bool> authState;
+
+  @override
+  List<Object> get props => [authState];
+}
