+++ packages/client/lib/app/modules/login/presentation/mobx/mobx.dart
@@ -1,0 +1,2 @@
+export 'main/main.dart';
+export 'coordinators/coordinators.dart';
